Updated navbar items, active labels
Added new navbar item titles for potential sections of the website (note: these sections are not necessarily set in stone, but rather are loose ideas of things that I may want to include on the website).

Also added some jQuery code that addas an "active" class to whatever navbar item is currently selected.