Included writeup on blog section, Google doc button
Basic framework of blog page in place; still need to decide whether it will just include listed out writings or has some sort of system to categorize the entries.

Button created that links to a "progress report" Google doc, which I think will be better to not have on the website due to its length.