Replaced projects.html with code.html
In effort to fix a bug where the projects.html page would load on the website without styling (this is caught by the loading screen on other pages), the shell of index.html was copy-pasted into code.html and filled with the contents of projects.html. This is unlikley to work, but I presently have few other ideas as to why the styling error is only happening on this one page.